+++ custom_components/econet300/number.py
@@ -16,6 +16,7 @@ from .const import (
     ENTITY_MAX_VALUE,
     ENTITY_MIN_VALUE,
     ENTITY_NUMBER_SENSOR_DEVICE_CLASS_MAP,
+    ENTITY_NUMBER_SENSOR_MODE_MAP,
     ENTITY_STEP,
     ENTITY_UNIT_MAP,
     NUMBER_MAP,
@@ -193,6 +194,7 @@ def create_number_entity_description(
         key=key,
         translation_key=camel_to_snake(map_key),
         device_class=ENTITY_NUMBER_SENSOR_DEVICE_CLASS_MAP.get(map_key),
+        mode=ENTITY_NUMBER_SENSOR_MODE_MAP.get(map_key),  # type: ignore[arg-type]
         native_unit_of_measurement=ENTITY_UNIT_MAP.get(map_key),
         native_min_value=min_value,
         native_max_value=max_value,